// src/api/generated/schemas/createClothesDto.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * StyleMind API
 * Documentación de la API de StyleMind
 * OpenAPI spec version: 1.0
 */

export interface CreateClothesDto {
  name: string;
  description: string;
  season: string;
  primaryColor: string;
  secondaryColor?: string;
  style: string;
  material?: string;
  size: string;
  /** Ids de las categorías relacionadas */
  categoriesId: string[];
}